Для MovieQuizViewControllerMock создан отдельный файл
