+++ MovieQuizPresenterTests/MovieQuizPresenterTests.swift
@@ -1,36 +1,6 @@
 import XCTest
 @testable import MovieQuiz
 
-final class MovieQuizViewControllerMock: MovieQuizViewControllerProtocol {
-    func setButtonsLocked(_ isLocked: Bool) {
-        
-    }
-    
-    func show(quiz step: QuizStepViewModel) {
-        
-    }
-    
-    func show(quiz result: QuizResultsViewModel) {
-        
-    }
-    
-    func highlightImageBorder(isCorrectAnswer: Bool) {
-        
-    }
-    
-    func showLoadingIndicator() {
-        
-    }
-    
-    func hideLoadingIndicator() {
-        
-    }
-    
-    func showNetworkError(message: String) {
-        
-    }
-}
-
 final class MovieQuizPresenterTests: XCTestCase {
     func testPresenterConvertModel() throws {
         let viewControllerMock = MovieQuizViewControllerMock()
