+++ MovieQuizPresenterTests/MovieQuizViewControllerMock.swift
@@ -1,0 +1,32 @@
+import Foundation
+@testable import MovieQuiz
+
+final class MovieQuizViewControllerMock: MovieQuizViewControllerProtocol {
+    func setButtonsLocked(_ isLocked: Bool) {
+        
+    }
+    
+    func show(quiz step: QuizStepViewModel) {
+        
+    }
+    
+    func show(quiz result: QuizResultsViewModel) {
+        
+    }
+    
+    func highlightImageBorder(isCorrectAnswer: Bool) {
+        
+    }
+    
+    func showLoadingIndicator() {
+        
+    }
+    
+    func hideLoadingIndicator() {
+        
+    }
+    
+    func showNetworkError(message: String) {
+        
+    }
+}
